+++ src/mass_driver/git.py
@@ -1,5 +1,6 @@
 """Manipulating git repos natively, without much knowledge of mass-driver models"""
 
+import logging
 from pathlib import Path
 from tempfile import mkdtemp
 
@@ -13,7 +14,7 @@ DEFAULT_CACHE = Path(".mass_driver/repos/")
 def clone_if_remote(repo_path: str, cache_folder: Path) -> GitRepo:
     """Build a GitRepo; If repo_path isn't a directory, clone it"""
     if Path(repo_path).is_dir():
-        print("Given an existing (local) repo: no cloning")
+        logging.info("Given an existing (local) repo: no cloning")
         # Clone it into cache anyway
         return GitRepo(path=repo_path)  # TODO: Actually clone-move the repo on the way.
     # SSH clone URL e.g: git@github.com:OverkillGuy/python-template
@@ -25,9 +26,9 @@ def clone_if_remote(repo_path: str, cache_folder: Path) -> GitRepo:
         repo_name = Path(repo_path).name
     clone_target = cache_folder / org / repo_name
     if clone_target.is_dir():
-        print("Given a URL for we cloned already: no cloning")
+        logging.info("Given a URL for we cloned already: no cloning")
         return GitRepo(clone_target)
-    print("Given a URL, cache miss: cloning")
+    logging.info("Given a URL, cache miss: cloning")
     cloned = GitRepo.clone_from(
         url=repo_path,
         to_path=clone_target,
@@ -41,7 +42,9 @@ def get_cache_folder(cache: bool) -> Path:
     cache_folder = DEFAULT_CACHE
     if not cache:
         cache_folder = Path(mkdtemp(suffix=".cache"))
-        print(f"Using repo cache folder: {cache_folder}/ (Won't wipe it on exit!)")
+        logging.info(
+            f"Using repo cache folder: {cache_folder}/ (Won't wipe it on exit!)"
+        )
     return cache_folder